Conflicts:
	DSPIC33E_hardware.h
	UI_page0.c
	main.c

+++ LCD_driver.c
@@ -38,7 +38,7 @@ static void WaitForFIFOUnLock(uint8 requestedMode);
 * Edit "count" based on processor speed for 1ms
 *
 *******************************************************************************/
-void delayms(int count)
+void delayms(uint16 count)
 {
     int i,j;
     for(i=0;i<count;i++)
@@ -173,7 +173,7 @@ void LCD_FrameAddr(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int 
 }
 
 /*******************************************************************************
-* Innitialize LCD Display
+* Initialize LCD Display
 *
 *
 *******************************************************************************/
