+++ LCD_driver.h
@@ -44,7 +44,7 @@ extern  uint16 POINT_COLOR;     // Pen color
 extern  uint8 num32_char[10];  // 32bit number to character array:
 
 // Functions
-void delayms(int count);          // 1ms delay routine used during display init
+void delayms(uint16 count);          // 1ms delay routine used during display init
 void Lcd_Init(void); 
 void LCD_Clear(uint16 Color);
